requirements.txt
python-telegram-bot==13.1
python-dotenv==0.19.2
psycopg2==2.9.3
python-dateutil==2.8.1
requests==2.25.0
websocket-client==1.4.1
beautifulsoup4==4.11.1